Move forking to a separate object
